Move redux files to properly folder
