+++ demo/redux/actions.js
