+++ demo/redux/reducers.js
